fix(status): share live proof-first truth across surfaces
